Split out cgo-dependent files with proper build tags
