+++ goTezosCrypto.go
@@ -1,3 +1,5 @@
+// +build cgo
+
 package goTezos
 
 import (
